chore: move background package from root to manager package
